fixed directory as required by the reviewer
