+++ README.md
@@ -1,117 +1,70 @@
+# Getting Started with Create React App
 
-# 📗 Table of Contents
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-- [📖 About the Project](#about-project)
-  - [🛠 Built With](#built-with)
-    - [Tech Stack](#tech-stack)
-    - [Key Features](#key-features)
-    - [🚀 Live Demo](#live-demo)
+## Available Scripts
 
-- [💻 Getting Started](#getting-started)
-  - [Prerequisites](#Prerequisites)
-- [👥 Authors](#authors)
-- [🔭 Future Features](#future-features)
-- [🤝 Contributing](#contributing)
-- [⭐️ Show your support](#support)
-- [🙏 Acknowledgements](#acknowledgements)
-- [📝 License](#license)
+In the project directory, you can run:
 
+### `npm start`
 
-# Math-Magicians
-A Mathematics Calculator built in React coupled with interesting randomly generated Mathematics quotes.
+Runs the app in the development mode.\
+Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
 
-## 🛠 Built With <a name="built-with"></a>
+The page will reload when you make changes.\
+You may also see any lint errors in the console.
 
-### Tech Stack <a name="tech-stack"></a>
+### `npm test`
 
-This project is built with React.
+Launches the test runner in the interactive watch mode.\
+See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
 
-<details>
-  <summary>REACT</summary>
-  <ul>
-    <li><a href="https://reactjs.org/">React.js</a></li>
-  </ul>
-</details>
+### `npm run build`
 
-### Key Features <a name="key-features"></a>
-- A classic Mathematics Calculator
-- Random Mathematics Quote Generator
-- A page for general information
+Builds the app for production to the `build` folder.\
+It correctly bundles React in production mode and optimizes the build for the best performance.
 
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
+The build is minified and the filenames include the hashes.\
+Your app is ready to be deployed!
 
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
 
-## 💻 Getting Started <a name="getting-started"></a>
-To get a local copy up and running, follow these steps.
-1. Clone this repository
-   run
-   `git clone https://github.com/arnoldnekemiah/Math-magicians.git`
+### `npm run eject`
 
-2. Open the project in a text editor
-3. Open your terminal and run `npm install`
-4. Run `npm start` to serve the project in your local browser
+**Note: this is a one-way operation. Once you `eject`, you can't go back!**
 
-### Prerequisites
+If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
 
-In order to run this project you need:
+Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
 
-1. Install git on your local computer
-> Find git installation details here https://git-scm.com/book/en/v2/Getting-Started-Installing-Git
+You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
 
-2. Clone this repository
-> Run this command `git clone https://github.com/arnoldnekemiah/mathmagicians.git`
+## Learn More
 
-3. Follow the instructions in the Get Started section above
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
+You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
 
+To learn React, check out the [React documentation](https://reactjs.org/).
 
-## 👥 Authors <a name="authors"></a>
-👤 Ikonde (Arnold) Nekemiah
+### Code Splitting
 
-- GitHub: [@arnoldnekemiah](https://github.com/arnoldnekemiah)
-- Twitter: [@arnoldikonde](https://twitter.com/arnoldikonde)
-- LinkedIn : [Ikonde (Arnold) Nekemiah](https://www.linkedin.com/in/ikonde-nekemiah-b050621ab/)
+This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
 
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
+### Analyzing the Bundle Size
 
+This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
 
-## 🔭 Future Features <a name="future-features"></a>
+### Making a Progressive Web App
 
-- [ ] Mobile version of the project
-- [ ] Updating the U.I styles
-- [ ] Use and management of Asynchronous code
-- [ ] Add screen responsiveness
+This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
 
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
+### Advanced Configuration
 
+This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
 
-## 🤝 Contributing <a name="contributing"></a>
+### Deployment
 
-Contributions, issues, and feature requests are welcome!
+This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
 
-Feel free to check the [issues page](https://github.com/arnoldnekemiah/Math-magicians/issues/). Open a New Issue and I'll look intently into your suggestions.
+### `npm run build` fails to minify
 
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-
-## ⭐️ Show your support <a name="support"></a>
-
-If you like this project **follow me on github** for more amazing projects
-and ideas
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-
-## 🙏 Acknowledgments <a name="acknowledgements"></a>
-
-I would like to thank **microverse** for the definition of the project requirements and provision of this template
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
-
-
-## 📝 License <a name="license"></a>
-
-This project is [MIT](./LICENSE) licensed.
-
-
-<p align="right">(<a href="#readme-top">back to top</a>)</p>
+This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
