--- src/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Math Magicians</title>
-</head>
-<body>
-  <div>
-    <h1>Hello World</h1>
-  </div>
-</body>
-</html>
